+++ Runtime/Async/PaletteMarkerProcessor.cs
@@ -41,52 +41,30 @@ public class PaletteMarkerProcessor : Yarn.Unity.ReplacementMarkupHandler
     /// <returns>A list of markup diagnostics if there are any errors, otherwise an empty list.</returns>
     public override List<LineParser.MarkupDiagnostic> ProcessReplacementMarker(MarkupAttribute marker, StringBuilder childBuilder, List<MarkupAttribute> childAttributes, string localeCode)
     {
-        // get the palette for this marker
-        if (!palette.FormatForMarker(marker.Name, out var style))
+        if (palette.PaletteForMarker(marker.Name, out var format))
         {
-            var error = new List<LineParser.MarkupDiagnostic>
+            childBuilder.Insert(0, format.Start);
+            childBuilder.Append(format.End);
+
+            // finally we need to know if we have to offset the markers
+            // most of the time we won't have to do anything
+            if (format.MarkerOffset != 0)
             {
-                new LineParser.MarkupDiagnostic($"Unable to identify a palette for {marker.Name}")
-            };
-            return error;
-        }
+                // we now need to move any children attributes down by however many characters were added to the front
+                // this is only the case if visible glyphs were added
+                // as in for example adding <b> to the front doesn't add any visible glyphs so won't need to offset anything
+                // and because markers are all 0-offset relative to parents
+                for (int i = 0; i < childAttributes.Count; i++)
+                {
+                    childAttributes[i] = childAttributes[i].Shift(format.MarkerOffset);
+                }
+            }
 
-        // do we have a custom colour set?
-        if (style.CustomColor)
-        {
-            childBuilder.Insert(0, $"<color=#{ColorUtility.ToHtmlStringRGBA(style.Color)}>");
-            childBuilder.Append("</color>");
+            return ReplacementMarkupHandler.NoDiagnostics;
         }
-
-        // do we need to bold it?
-        if (style.Boldened)
-        {
-            childBuilder.Insert(0, "<b>");
-            childBuilder.Append("</b>");
-        }
-        // do we need to italicise it?
-        if (style.Italicised)
-        {
-            childBuilder.Insert(0, "<i>");
-            childBuilder.Append("</i>");
-        }
-        // do we need to underline it?
-        if (style.Underlined)
-        {
-            childBuilder.Insert(0, "<u>");
-            childBuilder.Append("</u>");
-        }
-        // do we need to strikethrough it?
-        if (style.Strikedthrough)
-        {
-            childBuilder.Insert(0, "<s>");
-            childBuilder.Append("</s>");
-        }
-
-        // we don't need to modify the children attributes because TMP knows
-        // that the <color> tags aren't visible so we can just say we are done
-        // now
-        return ReplacementMarkupHandler.NoDiagnostics;
+        
+        var diagnostic = new LineParser.MarkupDiagnostic($"was unable to find a matching sprite for {marker.Name}");
+        return new List<LineParser.MarkupDiagnostic>() { diagnostic };
     }
 
     /// <summary>
@@ -100,7 +78,7 @@ public class PaletteMarkerProcessor : Yarn.Unity.ReplacementMarkupHandler
             return;
         }
 
-        if (palette.FormatMarkers.Count == 0)
+        if (palette.BasicMarkers.Count == 0)
         {
             return;
         }
@@ -110,7 +88,11 @@ public class PaletteMarkerProcessor : Yarn.Unity.ReplacementMarkupHandler
             lineProvider = (LineProviderBehaviour)GameObject.FindAnyObjectByType<DialogueRunner>().LineProvider;
         }
 
-        foreach (var marker in palette.FormatMarkers)
+        foreach (var marker in palette.BasicMarkers)
+        {
+            lineProvider.RegisterMarkerProcessor(marker.Marker, this);
+        }
+        foreach (var marker in palette.CustomMarkers)
         {
             lineProvider.RegisterMarkerProcessor(marker.Marker, this);
         }