+++ Runtime/Async/StyleMarkerProcessor.cs
@@ -1,4 +1,3 @@
-using System.Collections;
 using System.Collections.Generic;
 using System.Text;
 using UnityEngine;
@@ -20,7 +19,6 @@ namespace Yarn.Unity
         /// <inheritdoc/>
         public override List<LineParser.MarkupDiagnostic> ProcessReplacementMarker(MarkupAttribute marker, StringBuilder childBuilder, List<MarkupAttribute> childAttributes, string localeCode)
         {
-            // throw new System.NotImplementedException();
             // ok so we check if we have a property called style
             // if not give up
             if (!marker.TryGetProperty("style", out var property))
@@ -35,6 +33,10 @@ namespace Yarn.Unity
             childBuilder.Insert(0, $"<style=\"{property.StringValue}\">");
             childBuilder.Append("</style>");
 
+            // at this point we have no errors
+            // but it is entirely possible that style has added visible characters
+            // we have no way of knowing this
+            // so if this is the case any attributes will now be off
             return ReplacementMarkupHandler.NoDiagnostics;
         }
 
@@ -46,6 +48,11 @@ namespace Yarn.Unity
                 lineProvider = (LineProviderBehaviour)GameObject.FindAnyObjectByType<DialogueRunner>().LineProvider;
             }
             lineProvider.RegisterMarkerProcessor("style", this);
+
+            // in an ideal world instead of making you write out [style = h1] you could just do [h1]
+            // but TMP doesn't allow access to the list of styles, and as such also can't get their names
+            // so we have no way of know what the names of any style to register them
+            // alas
         }
     }
 }